% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flip.R
\name{flip}
\alias{flip}
\title{FLip the columns and row of a dataframe}
\usage{
flip(data)
}
\arguments{
\item{data}{A dataframe}
}
\value{
a flipped dataframe
}
\description{
FLip the columns and row of a dataframe
}
\examples{
flip(layout)
}
