% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removecolumn.R
\name{removecolumn}
\alias{removecolumn}
\title{Remove a column from a dataframe}
\usage{
removecolumn(data, what)
}
\arguments{
\item{data}{A dataframe}

\item{what}{what column to remove from a dataframe}
}
\value{
a dataframe with a deleted column
}
\description{
Remove a column from a dataframe
}
\examples{
removecolumn(layout, "Drug")
}
