% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replacevalue.R
\name{replacevalue}
\alias{replacevalue}
\title{replace in a value in a column based on something in another column}
\usage{
replacevalue(data, column1, firstvalue, replacement)
}
\arguments{
\item{data}{A dataframe}

\item{column1}{in what column do u want to replace a value}

\item{firstvalue}{what value do u want to replace}

\item{replacement}{what do u want to replace this value with}
}
\value{
a dataframe where the values of a specific column have been replaced by the user defined value
}
\description{
replace in a value in a column based on something in another column
}
\examples{
replacevalue(layout, 'Drug', "Analyfyn", "test")
}
