% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{subset}
\alias{subset}
\title{subset a dataframe}
\usage{
subset(data, what)
}
\arguments{
\item{data}{A dataframe}

\item{what}{what row do u want to subset}
}
\value{
a subsetted dataframe based on what row u want to subset
}
\description{
subset a dataframe
}
\examples{
subset(layout, 8)
}
