% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replacevalue.R
\name{replacevalue}
\alias{replacevalue}
\title{replace in a value in a column based on something in another column}
\usage{
replacevalue(data, column1, data1, value)
}
\arguments{
\item{data}{A dataframe}
}
\value{
a dataframe with a deleted column
}
\description{
replace in a value in a column based on something in another column
}
\examples{

}
